requirements.txt
# Core dependencies
openai>=1.3.7
tiktoken>=0.5.1
pdfplumber>=0.10.2
python-dotenv>=1.0.0

# CLI & development
argparse
tqdm

# Testing
pytest>=7.4.0

# Typing support
typing-extensions>=4.8.0

# Optional (for YAML or advanced templating in the future)
# pyyaml>=6.0
# jinja2>=3.1.2
